fix types after merge
